Add moving the regions as an option

+++ main/main.go
@@ -13,32 +13,48 @@ func main() {
 	verbose := flag.Bool("v", false, "verbose - log everything")
 	mode := flag.String("mode", "perChunk", "The mode by which the inhabited time will be compared to options: \"perChunk\"/\"regionSum\"")
 	regionDir := flag.String("path", "", "The path of the directory with the .mca files that should be cleaned")
+	newPath := flag.String("newPath", "", "The path where to move the region files to")
 	minInhabitTime := flag.Int("minInhabitedTicks", 250, "The value that has to be passed so that a chunk will be seen as \"used\"")
 	flag.Parse()
 
+	println("Move Dir:", *newPath)
 	println("RegionDir:", *regionDir)
 	println("InhabitedMin:", *minInhabitTime)
 	println("Verbose:", *verbose)
 	println("Mode:", *mode)
 
-	if len(*regionDir) == 0 {
-		println("No path to the region files was given, see -h for flags and their usage")
+	if len(*newPath) != 0 && !exists(*newPath) {
+		log.Fatal("The path to move the .mca files to doesn't exist")
 		return
 	}
 
+	if len(*regionDir) == 0 {
+		log.Fatal("No path to the region files was given, see -h for flags and their usage")
+		return
+	} else {
+		absPath, err := filepath.Abs(*regionDir)
+		if err != nil {
+			log.Fatal(*regionDir, "is not a recognized path")
+			return
+		}
+		println(*regionDir)
+		regionDir = &absPath
+		println(absPath)
+	}
+
 	if !exists(*regionDir) {
-		println(*regionDir, " doesn't exist")
+		log.Fatal(*regionDir, "doesn't exist")
 		return
 	}
 
 	if !strings.EqualFold(*mode, "perChunk") && !strings.EqualFold(*mode, "regionSum") {
-		println(*mode, "is not a valid comparison mode, must be on of \"perChunk\" or \"regionSum\"")
+		log.Fatal(*mode, "is not a valid comparison mode, must be on of \"perChunk\" or \"regionSum\"")
 		return
 	}
 
-	err := process(*verbose, strings.EqualFold(*mode, "perChunk"), *regionDir, int64(*minInhabitTime))
+	err := process(*verbose, strings.EqualFold(*mode, "perChunk"), *regionDir, *newPath, int64(*minInhabitTime))
 	if err != nil {
-		println(err)
+		log.Fatal(err)
 		return
 	}
 }
@@ -55,14 +71,16 @@ func exists(path string) bool {
 	return false
 }
 
-func process(verbose bool, perChunkMode bool, path string, minTime int64) (err error) {
+func process(verbose bool, perChunkMode bool, path string, newPath string, minTime int64) (err error) {
+	moveRegions := len(newPath) != 0
+
 	regions, err := filepath.Glob(path + string(os.PathSeparator) + "r.*.*.mca")
 	if err != nil {
 		return err
 	}
 
 	if len(regions) == 0 {
-		println("Couldn't find any region files in", path)
+		log.Println("Couldn't find any region files in", path)
 		return nil
 	}
 
@@ -74,7 +92,7 @@ regionLoop:
 		currentRegion, err := region.Open(file)
 		if err != nil {
 			if verbose {
-				println("Couldn't open file at", file, "skipping..")
+				log.Println("Couldn't open file at", file, "skipping..")
 			}
 			continue
 		}
@@ -90,7 +108,7 @@ regionLoop:
 				data, err := currentRegion.ReadSector(x, z)
 				if err != nil {
 					if verbose {
-						println("Couldn't read sector at x:", x, "z:", z, "from", file, "- skipping region as used..")
+						log.Println("Couldn't read sector at x:", x, "z:", z, "from", file, "- skipping region as used..")
 					}
 					continue regionLoop
 				}
@@ -98,19 +116,19 @@ regionLoop:
 				var chunk Chunk
 				err = chunk.Load(data)
 				if err != nil {
-					println("Couldn't read chunk at x:", x, "z:", z, "from", file, "- skipping region as used..")
+					log.Println("Couldn't read chunk at x:", x, "z:", z, "from", file, "- skipping region as used..")
 				}
 
 				if perChunkMode {
 					if chunk.Level.InhabitedTime > minTime {
-						println("Chunk at x:", chunk.XPos, "y:", chunk.YPos, "z:", chunk.ZPos, "is", chunk.Level.InhabitedTime, "ticks old, skipping region as used..")
+						log.Println("Chunk at x:", chunk.XPos, "y:", chunk.YPos, "z:", chunk.ZPos, "is", chunk.Level.InhabitedTime, "ticks old, skipping region as used..")
 						continue regionLoop
 					}
 				} else {
 					regionSum += chunk.Level.InhabitedTime
 					if regionSum > minTime {
 						if verbose {
-							println(file, "has exceeded the minimum of inhabitant time, skipping region as used..")
+							log.Println(file, "has exceeded the minimum of inhabitant time, skipping region as used..")
 						}
 						continue regionLoop
 					}
@@ -119,12 +137,19 @@ regionLoop:
 		}
 
 		if verbose {
-			println("Deleting", file, "..")
+			log.Println("Handling", file, "..")
 		}
 
-		err = os.Remove(file)
-		if err != nil && verbose {
-			println("Couldn't delete", file)
+		if moveRegions {
+			err = os.Rename(file, filepath.Join(newPath, filepath.Base(file)))
+			if err != nil && verbose {
+				log.Println("Couldn't move", file)
+			}
+		} else {
+			err = os.Remove(file)
+			if err != nil && verbose {
+				log.Println("Couldn't delete", file)
+			}
 		}
 	}
